+++ src/main/java/prismatic/shards/stellarity/registry/StellarityVillagerTypes.java
@@ -1,5 +1,6 @@
 package prismatic.shards.stellarity.registry;
 
+import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
@@ -8,13 +9,13 @@ import net.minecraft.world.entity.npc.villager.VillagerType;
 import prismatic.shards.stellarity.Stellarity;
 
 public interface StellarityVillagerTypes {
-	ResourceKey<VillagerType> END = register("end");
+	Holder.Reference<VillagerType> END = register("end");
 
-	private static ResourceKey<VillagerType> register(String id) {
+	private static Holder.Reference<VillagerType> register(String id) {
 		ResourceKey<VillagerType> key = Stellarity.key(Registries.VILLAGER_TYPE, id);
 		//noinspection InstantiationOfUtilityClass
-		Registry.register(BuiltInRegistries.VILLAGER_TYPE, Stellarity.id(id), new VillagerType());
-		return key;
+		return Registry.registerForHolder(BuiltInRegistries.VILLAGER_TYPE, Stellarity.id(id), new VillagerType());
+
 	}
 
 	static void init() {